requirements.txt
pyautogui>=0.9.53
keyboard>=0.13.5
pandas>=1.3.0
pyperclip>=1.8.2
Pillow>=8.2.0
numpy>=1.21.0
openpyxl>=3.0.7 